Implement method for generate masks and plot them
Implemented a method to generate mask that generate only the mask for the concepts present on the image, and another method that generate a mask for all concepts, if one concept is not present it generates a mask with all False values. Then some function to plot images for testing purposes are added.